Class3/voice-agent/requirements.txt
fastapi
uvicorn[standard]
python-multipart
faster-whisper
ffmpeg-python
openai
python-dotenv
pyttsx3
